requirements.txt
Django
djangorestframework
pytest
django-admin
django-extensions
ipython
python-dotenv
drf-yasg
psycopg2-binary